refactor p1
