+++ assets/js/platformer/archive/PlayerMiniHogwarts.js
@@ -1,6 +1,6 @@
-import GameEnv from './GameEnv.js';
-import PlayerBase from './PlayerBase.js';
-import GameControl from './GameControl.js';
+import GameEnv from '../GameEnv.js';
+import PlayerBase from '../PlayerBase.js';
+import GameControl from '../GameControl.js';
 
 /**
  * @class PlayerHills class
